requirements.txt
streamlit
fastapi
statsbombpy
pandas
langchain-google-genai
langchain-tools
uvicorn
pydantic
plotly
seaborn
matplotlib